Overload methods, fix return value in literal initialisers
Add tests for literals that have values

+++ commit-based-cipm/fi-tests/cipm.consistency.fitests/src/cipm/consistency/fitests/similarity/java/initialiser/literals/IHexLongLiteralInitialiser.java
@@ -16,7 +16,7 @@ public interface IHexLongLiteralInitialiser extends ILongLiteralInitialiser {
     	return true;
     }
     
-    public default boolean setHexValue(HexLongLiteral hll, int val) {
+    public default boolean setHexValue(HexLongLiteral hll, long val) {
     	return this.setHexValue(hll, BigInteger.valueOf(val));
     }
 }
