+++ commit-based-cipm/fi-tests/cipm.consistency.fitests/src/cipm/consistency/fitests/similarity/java/initialiser/literals/IOctalLongLiteralInitialiser.java
@@ -16,7 +16,7 @@ public interface IOctalLongLiteralInitialiser extends ILongLiteralInitialiser {
     	return true;
     }
     
-    public default boolean setOctalValue(OctalLongLiteral oll, int val) {
+    public default boolean setOctalValue(OctalLongLiteral oll, long val) {
     	return this.setOctalValue(oll, BigInteger.valueOf(val));
     }
 }
